Confgen CLI moved into the python module.
